requirements.txt
# Core
langchain==0.1.0
langchain-community==0.0.10
streamlit==1.30.0

# LLMs
openai==1.10.0           

# Data & Analysis
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.1

# Database
pyodbc==5.0.1            
duckdb         
sqlalchemy==2.0.20   
  
# Tools
python-dotenv==1.0.0
pydantic==2.5.3

# Dev & Testing
pytest==7.4.4
black==23.12.1
flake8==7.0.0

# (Opcional) Avançado
chromadb==0.4.22         # Vector DB para memória
langsmith==0.0.77        # Observability (debugging agents)